--- src/useArraySetState.jsx
@@ -1,9 +1,0 @@
-export default function useArraySetState(setter) {
-  return (value, order) =>
-    setter((prevValue) => {
-      const newValue = [...prevValue];
-      newValue[order] = value;
-
-      return newValue;
-    });
-}
